Add a Job linter
Add a Linter that can catch obvious errors & ommisions in a Job definition. The core class
implements the bulk of the functionality and it is intended to be used by both the Client
to Lint a job locally and by the API to lint a job prior to dispatch and provide useful
information back to the user.